requirements.txt
langchain
langchain-openai
openai
python-dotenv
sqlalchemy
pydantic
pillow
pdf2image
pytesseract
pandas